fix:カード詳細画面をモーダルに変更
