--- ui/Popover/index.tsx
@@ -1,1 +1,0 @@
-export { StyledPopover } from "./StyledPopover";
